API-python/requirements.txt
flask
flask_restplus
mysql-connector-python
